page animation added
